Add file that trie to check for a compiler optimization
It tries to see if the compiler could optimize the "item &&"
part of the check out of the loop. It seems that both gcc and
clang were able to optimize at various optimization levels (possibly,
-O1 and -O2 respectively, though I'm not sure).

This was done for the 'git' mailing list. Searching for
"for_each_string_list_item" might find the relevant thread
which has more information.