(BUG FIX) fix issue with replace_end not doing replacement
The 'replace_end' function didn't do replace the last occurences of
the search terms in cases in which the search term was a character and
there were other instances of the character in the input. This is because
the conditional was wrong from the start in assuming that if the first
instance was not the last occurence then there will be "NO" last
occurences in the input and thus it returned the input as is back to the
caller.

This wasn't noticed before as it was used only for "word searching". The
issue showed it's ugly head when trying to replace characters.

Fix the issue by moving the check for last occurence to the correct location
and handling the alternate case correctly.