Remove instances of %m
It's not ISO C. It would have been better if 'fgets' didn't add
the line terminator while reading input as it's more secure in other
ways.

Unfortunately, I'm left with using plain scanf (which is not so secure)
as 'fgets' adds the line terminator to input; until I find a better way
that's more useful. (Fortunately, this is not going to be used by anyone
in the wild, I hope !!)