Add implementation that sees where an array is allocated memory (?)
This was done as there was a little misconception that it is
"wrong" to pass the address of a local variable to another
function (which will be proved wrong by a future test).

This implementation tries to see various addresses and deduce
where it's allocated memory. Of course it's allocated memory in
the stack as should be the case

The other outcome of this test is that the memory of consecutively
declared variables reduce pointing that the stack grows from the
bottom towards the top end of memory.

With `gcc where_is_array_stored.c`:

    Address of a: 0x7ffd8b987c78

    Address IN arr: 0x7ffd8b987c50
    Address OF arr: 0x7ffd8b987c50

    Address of b: 0x7ffd8b987c48

Of course kicking in optimization gives different results, so the
order in which varaibles are declared is NOT something that we
could use to deduce the address of a variable.

With `gcc where_is_array_stored.c -O3`:

    Address of a: 0x7fffd1c38010

    Address IN arr: 0x7fffd1c38020
    Address OF arr: 0x7fffd1c38020

    Address of b: 0x7fffd1c38018