Add a test implementation that demonstrates "undefined overflows"
When we guarantee GCC that the code doesn't have any arithmetic that
overflows, it starts treating overflowing arithmetic as 'undefined'
and triggers some optimisations which can be done when arithmetics
are guaranteed not to overflow.

Demonstrate the consequence that occurs when arithmetic overflows
when optimisation triggers (as GCC has been guaranteed that arithmetic
doesn't overflow).