intern/sudoku: fix a grave copy-paste BUG
Thiis bug was a consequence of copy pasting and making slight changes
to re-use code instead of moving it into a function and avoiding
repetition.

This teaches the (possibly well-known) lesson that you should think a
little and move the common piece of code into a function rather than
just repeating it and making slight changes to it. DOT.