intern/sudoku: refactor the code
This refactor avoids the "grave copy/paste bugs" that have been observed
so far. It's better to do something like the consequence of the change
than to copy and paste and modify the code.

It also avoids a function from returning the possibility as it seems
cumbersome and bug-prone. It's easy to update the value directly when
required.