intern/sudoku_solver: correctly condition compile
As a consequence of 0c66207 (intern/sudoku_solver: fix (correctly
order) debug output, 2018-02-14), a required function call went into
a conditional compilation block which shouldn't not have been done
as the block was only meant for debug output and not for side effects.

So, move the statement out of the conditional compilation block to
ensure it is always compiled.