intern/sudoku_solver: avoid adjusting possibility of newly filled cells
Only the zero-check is used to cells yet to be filled. So, there's no use
in updating the possibility of newly filled cells. Further, it would not
reflect the state of the possibility vector correctly unless they are updated.
Updating the possibility vector is not of much use either.

So, ...