intern/sudoku_solver: add initial version of "naked double" detector
This idenitfies naked doubles found in rows and prints them when debug
output is enabled.

This is done using a brute-force method rather than trying to generate
subsets of cells in a group and checking for possible "naked" doubles,
triples etc. due to lack of time (and lack of necessity [Necessity is
the mother of invention!] ;-).