intern/sudoku_solver: complete "naked double" detection
This implementation extends the intial version and correctly identifies
naked doubles in different groups using a brute force technique. This
results in some "hard" puzzles getting solved and some getting a partial
solution.

As there is a possibility of detecting the same naked double again and
again the number of rounds is limited to 100.