intern/maze: avoid special cases in adjacency list initialization
This is done by taking advantage of the fact that only the start and end
gate pixels are open in the circumference of the maze.

So, instead of special casing the gate pixels inside the loop, reduce the
range over which the loop runs and add the adjacency of the gate pixels
outside the loop.