intern/maze: boundaries may contain clear pixels other than start/end gates
The assumption that only the start and end gates are the only clear
pixels in the circumference doesn't seem to hold for all images. This
seems to be causing issues while pruning.

So, reduce the assumption to "The only clear pixels in the first and
last row are the start and end gate pixels correspondingly".