intern/maze: test if this speeds up code
ae05e1a (Revert "intern/maze: inline some bmp related code to gain it
speeds up", 2018-02-27) mentions about an alternative to the speed up it
took back for the sake of abstraction. This commit attempts at the
mentioned speed up.

The observation was that the speed up is the same as that achieved by
877c45f (intern/maze: optimize some code to gain some speed up,
2018-02-27)

Some sample timings on a Quad-core CPU ticking at 3GHz with 4GB RAM,

	Image dimensions (in pixels):
	width: 4001	 height: 4001
	Size of image data: 48028004
	Successfully found the shortest path between from the source to the
	destination.
	The solution has been written to the file.

	Before speedup:

		real	0m2.430s
		user	0m2.172s
		sys	0m0.256s

	Speedup of 877c45f:

		real	0m2.370s
		user	0m2.116s
		sys	0m0.252s

	Speedup of this commit:

		None (Same as that before speedup)