intern/maze: switch to the A-star algorithm from BFS
For the single source single destination problem, the A-star algorithm
is known to search for the path towards the destination given the
"correct heuristic". This is the best algorithm for the finding the
shortest path from the source to destination in a maze.

The number of nodes searched by A-star is less than that searched by BFS
for the "manhattan distance" heuristic. This does not mean the running
time for A-star is less than that of BFS. This is (possibly) because A-star
uses a priority queue while BFS uses only a simple queue.