intern/maze: optimise graph creation and initialization
This update merges the adjacency list initialization and heuristic value
initialization into the create_graph function. This avoids repeatedly
doing separate computation when they could be combined without issues.

This optimisation is based on the observations:

 - Space wasted in structure padding could be used to store the
   heuristic value:

   The node structure wastes a lot of space in terms of structure padding.
   So, adding the heuristic value into the node doesn't increase the memory
   requirements for the node but rather decreases the space wasted in terms
   of structure padding.

 - Adjacency could be initialized while creating the graph:

   The optimized adjacency initialization [537c078 (intern/maze: optimize
   adjacency list initialization, 2018-02-23] allows us to initalize the
   adjacency for the nodes while creating the graph.

This speeds up the program a lot for huge inputs and reduces the memory
requirement a little.