list: properly return from the insert function
Not returning from the insert function results garbage being
returned in some cases as a consequence of which sometime negative
values are returned indicating error when the function did not
fail actually.

Correct this mistake by properly returning success from the function.

Tip: It's better to use -Wall flag which could have caught this sooner.